+++ sdk/src/constants/txConstants.ts
@@ -1,1 +1,6 @@
+import { PublicKey } from '@solana/web3.js';
+
 export const NOT_CONFIRMED_ERROR_CODE = -1001;
+export const FUEL_RESET_LOG_ACCOUNT = new PublicKey(
+	'FuE1gqp2fzw2sDNLrbZqKsqpphJcoSW6HPaSJjGd4RZ4'
+);